fix: changed data type for Project from int to Long.

+++ src/main/java/com/construtech/buildsphere/platform/resourceManagement/interfaces/rest/resources/MachineResource.java
@@ -9,6 +9,6 @@ public record MachineResource(
         LocalDate receptionDate,
         LocalDate endDate,
         double totalCost,
-        int projectId
+        Long projectId
 ) {
 }
